damn it

+++ cppRef/Data Structures/Graph/Edge.cpp
@@ -1,6 +1,6 @@
 #include "Edge.h"
 
-Edge::Edge(GraphNode& node1, GraphNode& node2, int weight) : node1(node1), node2(node2), weight(weight) {
+Edge::Edge(GraphNode& node1, GraphNode& node2, double weight) : node1(node1), node2(node2), weight(weight) {
     node1.connect(*this);
     node2.connect(*this);
 }
